Implement Action instead of full Subscriber
Just as we often do not want to implement a full OnSubscribe object
to send in using the Observable.create method often we also do not
need to implement a full Subscriber object. We might not care about
all the callback methods in the Subscriber. Instead, as here, we
might only care about the onNext method. By creating an Action1
object we can implement what should happen in the onNext method
inside the Action1 objects call method and then just send the
Action1 object in to the Observable using a variant of the
subscribe method.