Using the map operator
Now comes some of the cool stuff with RxJava. Using Observables
Subscribers is basically just a listener (Observer) pattern modifies
to work delivering a stream of objects. But now we start looking at
some more fancy stuff. Using the map operator (method) we can
modify the objects that are emited by the Observable before they
get delivered to the Subscribers. In this case we use the Func1
object which wraps a method with one inparameter. Using this we
can makes sure that an extra String is added to the emitted String.