Some more use of chained map operators
Since the map methods can be chained we can go wild!
In this commit we first add the extra String to the end
of the String provided by the Observable but then we use
other map methods to first get the hash of the new String
and then convert the Integer hash to a String. This also
shows that maps can freely convert between different types
of objects so that we are able to transform a whole stream
of objects from one type to another.