Using flatmap to flatten collections
When we get an Observable that is a Collection of some sort and
we want to transform the items in the collections the way we
usually do with a map etc we can use the flatmap operator (method)
to transform the Observable returned to the Subscriber from a
Collection object into the individual items in the collection.
This means that we can then apply all our usual transforms on the
individual items. In this case we prepend a String before the String
returned from the Observable.