Further use of flatmap
Actually we do not need to use flatmap only when operating
on collections of objects. The main difference between flatmap
and map is that flatmap returns an Observable whereas map returns
an Object that is then wrapped in an Observable. This means that
we can use flatmap in situations like this one where we have a
method that returns an Observable. By using flatmap we can directly
return the result of the method. Also this commit moves the RxJava
code to be run when pressing the floating action button.