Handling of errors
We implement a full subscriber with the onCompleted and onError
methods implemented. This allows us to capture if an error occurs
in a simple way. The error handling code can reside only with the
subscriber. Also we can clearly see whether we have received all
the items by the onComplete method being called.