JavaRx threading
By using the subscribeOn operator we can specify which thread all
subscriptions should run on. This will affect the whole chain of
operators that operate on the items emitted. The emitting is however
still done on the thread it is started from. We can also use the
observeOn operator. It works a bit differently than the subscribeOn
operator in that it will only affect operators that come AFTER it.
This example also shows how we can specify and use our own thread
pools by using the Schedulers.from method to specify the thread to
use. We have also added the rxandroid library in order to use the
AndroidSchedulers.mainThread() Scheduler to do work on the main
thread.